refactor: move events to their own files
