+++ src/main/java/com/stefancooper/SpigotUHC/Plugin.java
@@ -7,6 +7,8 @@ import com.stefancooper.SpigotUHC.commands.ResumeCommand;
 import com.stefancooper.SpigotUHC.commands.SetConfigCommand;
 import com.stefancooper.SpigotUHC.commands.StartCommand;
 import com.stefancooper.SpigotUHC.commands.ViewConfigCommand;
+import com.stefancooper.SpigotUHC.events.BaseEvents;
+import com.stefancooper.SpigotUHC.events.ReviveEvents;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -24,7 +26,8 @@ public class Plugin extends JavaPlugin implements Listener {
     public void onEnable() {
         config = new Config(this);
         Defaults.setDefaultGameRules(this.config);
-        Bukkit.getPluginManager().registerEvents(new Events(config), this);
+        Bukkit.getPluginManager().registerEvents(new BaseEvents(config), this);
+        Bukkit.getPluginManager().registerEvents(new ReviveEvents(config), this);
         started = false;
         System.out.println("UHC Plugin enabled");
     }
